+++ src/rngs/mod.rs
@@ -95,15 +95,15 @@ mod xoshiro256plusplus;
 
 #[cfg(feature = "std_rng")]
 mod std;
-#[cfg(all(feature = "std", feature = "std_rng", feature = "getrandom"))]
+#[cfg(all(feature = "std", feature = "std_rng", feature = "os_rng"))]
 pub(crate) mod thread;
 
 #[cfg(feature = "small_rng")]
 pub use self::small::SmallRng;
 #[cfg(feature = "std_rng")]
 pub use self::std::StdRng;
-#[cfg(all(feature = "std", feature = "std_rng", feature = "getrandom"))]
+#[cfg(all(feature = "std", feature = "std_rng", feature = "os_rng"))]
 pub use self::thread::ThreadRng;
 
-#[cfg(feature = "getrandom")]
+#[cfg(feature = "os_rng")]
 pub use rand_core::OsRng;